requirement.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-amazon
apache-airflow-providers-postgres
pyspark==3.5.4
boto3
s3fs
pandas
numpy
sqlalchemy
psycopg2-binary
pyarrow